refactored: database
