+++ srcs/ai/utils/database/connection/__init__.py
@@ -1,0 +1,1 @@
+from .client import client, DB_NAME
